examples-lib (#1092)
* **examples-lib**: new shared package for examples
* Refactor: Sstart migrating code from chat* into examples-lib
