+++ ts/examples/examplesLib/src/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation and Henry Lucco.
+// Licensed under the MIT License.
+
+export * from "./chalkWriter.js";
