+++ ts/examples/examplesLib/src/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "compilerOptions": {
+    "composite": true,
+    "rootDir": ".",
+    "outDir": "../dist"
+  },
+  "ts-node": {
+    "esm": true
+  }
+}
