+++ ts/examples/examplesLib/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "composite": true,
+  },
+  "include": [],
+  "references": [
+    { "path": "./src" },
+  ],
+  "ts-node": {
+    "esm": true
+  }
+}
